Fixed a Date Parsing Issue. Major other issues.
There are some major issues with how messages are downloaded. Messages
are always appended even if they are duplicates. Probably need an ID
attribute or maybe use a hash of the text and the date to check for
duplicates.

Also, need to figure out the classic tableviewcell problem of making
sure asynchronous tasks get taken care of applied to the right cell.
Right now, there are checks to make sure the responding download
callback is for that cell and the result is thrown out if its not. But
if that happens then things don’t load correctly.

Some caching would go a long way, but that won’t solve it on its own.