Updated Download Manager
The download manager now knows when all connections have succeeded or
failed and doesn’t update the data source until all connections are
done. This also eliminates duplicates. When the data source is updated
with the new array. It posts a notification to let any object know it
may want to refresh the view.