Refactored for Value Types
All structs have been refactored as 100% value types. Now all the
custom data is is value types (as opposed to Object types) and they all
conform to several protocols. They are guaranteed to make valid NSDates
and NSURLs when requested. They also can be written to and read from
NSUserDefaults and NSJSONSerialization with the Writable protocol.