Bunch of small changes
I created a bunch of fake data subscriptions (not part of the code) to
see how performance was going. Downloading and creating messages is
extremely fast. Even when there are 100 or so messages there is no
pause when downloading them or converting those downloads into messages.

Also, once they are in the table, the tableview scrolls really well.
Also, the error checking works well. It can detect issues where the
download didn’t work (404, etc) it can also detect issues where there
was unreadable data in the file.

The part that is super slow is sorting all the messages by date. This
happens in the view controller when it loads. I’m using Swift’s .sort
mechanism and its pretty slow. Its probably also not helped by using
NSDate and all the NSDates are created dynamically by the messages on
demand :-/

Lastly, I refactored all the dictionaries to use String types instead
of NSURL. Since we are looping through things, the String type doesn’t
have to be generated dynamically in tight loops. It also means we get
to pass around only Value types and not object types.