Added a New NilCoalescing Operator for more safety
Found a new way to use ?? for safety and to provide default values, but
still know if there is something wrong. The !! operator crashes when in
debug build and the optional is nil. This lets you know the default
value provided by ?? will always be used, which is probably not what
you intended.